feat: add more types

+++ index.ts
@@ -1,33 +1,45 @@
 import express, { Request, Response, NextFunction } from 'express';
 import multer from 'multer';
 import { exec } from "child_process";
-import { Database } from 'sqlite3';
+import sqlite3 from 'sqlite3';
 import cors from 'cors';
-import jwt from 'jsonwebtoken';
+import jwt, { JwtPayload } from 'jsonwebtoken';
 import bcrypt from 'bcrypt';
+
+interface User {
+    username: string;
+    password: string;
+}
+
+interface RequestWithUser extends Request {
+    user?: string | JwtPayload;
+}
 
 const app = express();
 const upload = multer({ dest: "uploads/" });
 
-let db = new Database('./db.sqlite3', (err) => {
-    if (err) {
-        console.error(err.message);
-    }
-    console.log('Connected to the sqlite database.');
-    db.run('CREATE TABLE IF NOT EXISTS users(username TEXT UNIQUE, password TEXT)', (err) => {
+let db: sqlite3.Database;
+
+function openDb() {
+    db = new sqlite3.Database('./db.sqlite3', (err) => {
         if (err) {
-            console.error(err.message);
+            console.error(`Could not open database: ${err.message}`);
+            process.exit(1);
         }
     });
-});
+
+    db.run('CREATE TABLE IF NOT EXISTS users(username TEXT UNIQUE, password TEXT)');
+}
+
+openDb();
 
 app.use(cors());
 app.use(express.json());
 
-app.post('/login', async (req, res) => {
-    const { username, password } = req.body;
+app.post('/login', async (req: Request, res: Response) => {
+    const { username, password } = req.body as User;
 
-    db.get("SELECT * FROM users WHERE username = ?", [username], async (err, row) => {
+    db.get("SELECT * FROM users WHERE username = ?", [username], async (err, row: User) => {
         if (err) {
             return console.error(err.message);
         }
@@ -45,7 +57,7 @@ app.post('/login', async (req, res) => {
     });
 });
 
-const verify = (req: Request, res: Response, next: NextFunction) => {
+const verify = (req: RequestWithUser, res: Response, next: NextFunction) => {
     const token = req.header('auth-token');
     if (!token) return res.status(401).json('Access Denied');
 
@@ -60,17 +72,19 @@ const verify = (req: Request, res: Response, next: NextFunction) => {
 
 app.use("/packages", express.static("packages"));
 
-app.post("/upload", [verify, upload.single("package")], (req, res) => {
-    exec(`mv ${req.file.path} packages/${req.file.originalname} && rm -r uploads/*`, (error) => {
-        if (error) {
-            console.error(`Error moving file: ${error}`);
-            res.sendStatus(500);
-        } else {
-            res.sendStatus(200);
-        }
-    });
+app.post("/upload", [verify, upload.single("package")], (req: RequestWithUser, res: Response) => {
+    if (req.file && "originalname" in req.file) {
+        exec(`mv ${req.file.path} packages/${req.file.originalname} && rm -r uploads/*`, (error) => {
+            if (error) {
+                console.error(`Error moving file: ${error}`);
+                res.sendStatus(500);
+            } else {
+                res.sendStatus(200);
+            }
+        });
+    }
 });
 
-app.listen(5000, () => {
-    console.log("Server started on port 5000");
+app.listen(4269, () => {
+    console.log("Server started on port 4269");
 });
